fix: cursor comments
