+++ app/components/UI/Predict/views/PredictBuyWithAnyToken/hooks/usePredictBuyInfo.ts
@@ -4,6 +4,7 @@ import { useTransactionPayTotals } from '../../../../../Views/confirmations/hook
 import { usePredictPaymentToken } from '../../../hooks/usePredictPaymentToken';
 import { OrderPreview } from '../../../types';
 import { useInsufficientPayTokenBalanceAlert } from '../../../../../Views/confirmations/hooks/alerts/useInsufficientPayTokenBalanceAlert';
+import { useNoPayTokenQuotesAlert } from '../../../../../Views/confirmations/hooks/alerts/useNoPayTokenQuotesAlert';
 
 interface UsePredictBuyInfoParams {
   currentValue: number;
@@ -23,8 +24,21 @@ export const usePredictBuyInfo = ({
   const { isPredictBalanceSelected } = usePredictPaymentToken();
   const payTotals = useTransactionPayTotals();
 
-  const [insufficientPayTokenBalanceAlert] =
-    useInsufficientPayTokenBalanceAlert();
+  const insufficientPayAlerts = useInsufficientPayTokenBalanceAlert();
+  const noQuotesAlerts = useNoPayTokenQuotesAlert();
+
+  const blockingPayAlerts = useMemo(() => {
+    const allPayAlerts = [...insufficientPayAlerts, ...noQuotesAlerts];
+    return allPayAlerts.filter((a) => a.isBlocking);
+  }, [insufficientPayAlerts, noQuotesAlerts]);
+
+  const hasBlockingPayAlerts =
+    !isPredictBalanceSelected && blockingPayAlerts.length > 0;
+
+  const blockingPayAlertMessage = useMemo(
+    () => blockingPayAlerts[0]?.message ?? blockingPayAlerts[0]?.title,
+    [blockingPayAlerts],
+  );
 
   const [acceptedDepositFee, setAcceptedDepositFee] = useState(0);
 
@@ -37,22 +51,14 @@ export const usePredictBuyInfo = ({
   );
 
   const computedDepositFee = useMemo(() => {
-    if (
-      isPredictBalanceSelected ||
-      !payTotals?.fees ||
-      insufficientPayTokenBalanceAlert
-    )
+    if (isPredictBalanceSelected || !payTotals?.fees || hasBlockingPayAlerts)
       return 0;
     const { provider, sourceNetwork, targetNetwork } = payTotals.fees;
     return new BigNumber(provider?.usd ?? 0)
       .plus(sourceNetwork?.estimate?.usd ?? 0)
       .plus(targetNetwork?.usd ?? 0)
       .toNumber();
-  }, [
-    insufficientPayTokenBalanceAlert,
-    isPredictBalanceSelected,
-    payTotals?.fees,
-  ]);
+  }, [isPredictBalanceSelected, payTotals?.fees, hasBlockingPayAlerts]);
 
   useEffect(() => {
     if (computedDepositFee > 0) {
@@ -99,5 +105,7 @@ export const usePredictBuyInfo = ({
     total,
     rewardsFeeAmount,
     totalPayForPredictBalance,
+    blockingPayAlertMessage,
+    hasBlockingPayAlerts,
   };
 };
